% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{acres_to_hectares}
\alias{acres_to_hectares}
\title{Convert Acres to Hectares}
\usage{
acres_to_hectares(x)
}
\arguments{
\item{x}{Numeric vector in acres.}
}
\value{
Numeric vector in hectares.
}
\description{
Convert Acres to Hectares
}
\keyword{internal}
