% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ba_ft2}
\alias{ba_ft2}
\title{Basal Area in Square Feet}
\usage{
ba_ft2(dbh_inches)
}
\arguments{
\item{dbh_inches}{Numeric vector of diameters at breast height in inches.}
}
\value{
Numeric vector of basal area in square feet.
}
\description{
Computes individual tree basal area from DBH in inches.
}
\examples{
ba_ft2(10)   # 0.5454 ft^2
ba_ft2(c(5, 10, 15))

}
