% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ba_m2}
\alias{ba_m2}
\title{Basal Area in Square Meters}
\usage{
ba_m2(dbh_cm)
}
\arguments{
\item{dbh_cm}{Numeric vector of diameters at breast height in centimeters.}
}
\value{
Numeric vector of basal area in square meters.
}
\description{
Computes individual tree basal area from DBH in centimeters.
}
\examples{
ba_m2(25.4)

}
