% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cm_to_inches}
\alias{cm_to_inches}
\title{Convert Centimeters to Inches}
\usage{
cm_to_inches(x)
}
\arguments{
\item{x}{Numeric vector in centimeters.}
}
\value{
Numeric vector in inches.
}
\description{
Convert Centimeters to Inches
}
\keyword{internal}
