% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{.compute_top_height}
\alias{.compute_top_height}
\title{Compute top height (mean height of 40 largest TPA trees)}
\usage{
.compute_top_height(height, tpa, dbh)
}
\description{
Compute top height (mean height of 40 largest TPA trees)
}
\keyword{internal}
