% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{.filter_fia_by_plot_cn}
\alias{.filter_fia_by_plot_cn}
\title{Filter all FIA tables to selected plot CNs}
\usage{
.filter_fia_by_plot_cn(fia_data, plot_cns)
}
\description{
Filter all FIA tables to selected plot CNs
}
\keyword{internal}
