% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_codes.R
\name{.load_variant_csv}
\alias{.load_variant_csv}
\title{Load Variant-Specific CSV}
\usage{
.load_variant_csv(variant)
}
\arguments{
\item{variant}{FVS variant code.}
}
\value{
Data frame or NULL if file not found.
}
\description{
Load Variant-Specific CSV
}
\keyword{internal}
