% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{.spatial_filter_python}
\alias{.spatial_filter_python}
\title{Python-based spatial filter (fallback)}
\usage{
.spatial_filter_python(aoi, plots, fia_data, buffer_ft)
}
\description{
Python-based spatial filter (fallback)
}
\keyword{internal}
