% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{feet_to_meters}
\alias{feet_to_meters}
\title{Convert Feet to Meters}
\usage{
feet_to_meters(x)
}
\arguments{
\item{x}{Numeric vector in feet.}
}
\value{
Numeric vector in meters.
}
\description{
Convert Feet to Meters
}
\keyword{internal}
