% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_codes.R
\name{get_common_name}
\alias{get_common_name}
\title{Get Common Name from FIA SPCD}
\usage{
get_common_name(spcd)
}
\arguments{
\item{spcd}{Integer vector of FIA SPCD values.}
}
\value{
Character vector of common names. Unknown species return
"unknown species".
}
\description{
Returns common tree names for FIA species codes.
}
\examples{
get_common_name(c(833, 12, 129))
# "northern red oak", "balsam fir", "eastern white pine"

}
\seealso{
Other Species Lookups: 
\code{\link{fia_to_fvs}()},
\code{\link{fvs_to_fia}()},
\code{\link{get_species_table}()},
\code{\link{pef_to_fvs}()}
}
\concept{Species Lookups}
