% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{get_remeasured_plots}
\alias{get_remeasured_plots}
\title{Find Remeasured Plots}
\usage{
get_remeasured_plots(fia_data)
}
\arguments{
\item{fia_data}{List of FIA data frames (must contain PLOT table).}
}
\value{
A tibble with columns:
\describe{
\item{plot_cn}{Plot control number (most recent).}
\item{state_code, county_code}{Location.}
\item{n_measurements}{Number of times the plot was measured.}
\item{first_year, last_year}{Range of inventory years.}
\item{years}{Comma-separated list of measurement years.}
}
}
\description{
Identifies FIA plots that have been measured in multiple inventory cycles,
essential for growth analysis and change detection.
}
\examples{
\dontrun{
fia <- fetch_fia_data("ME")
remeas <- get_remeasured_plots(fia)
remeas[remeas$n_measurements >= 3, ]
}

}
\seealso{
\code{\link[=build_time_series]{build_time_series()}}, \code{\link[=summarize_growth]{summarize_growth()}}
}
