% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hectares_to_acres}
\alias{hectares_to_acres}
\title{Convert Hectares to Acres}
\usage{
hectares_to_acres(x)
}
\arguments{
\item{x}{Numeric vector in hectares.}
}
\value{
Numeric vector in acres.
}
\description{
Convert Hectares to Acres
}
\keyword{internal}
