% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inches_to_cm}
\alias{inches_to_cm}
\title{Convert Inches to Centimeters}
\usage{
inches_to_cm(x)
}
\arguments{
\item{x}{Numeric vector in inches.}
}
\value{
Numeric vector in centimeters.
}
\description{
Convert Inches to Centimeters
}
\keyword{internal}
