% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{meters_to_feet}
\alias{meters_to_feet}
\title{Convert Meters to Feet}
\usage{
meters_to_feet(x)
}
\arguments{
\item{x}{Numeric vector in meters.}
}
\value{
Numeric vector in feet.
}
\description{
Convert Meters to Feet
}
\keyword{internal}
