% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_codes.R
\name{pef_to_fvs}
\alias{pef_to_fvs}
\title{Convert PEF Species Code to FVS Species Code}
\usage{
pef_to_fvs(pef_code)
}
\arguments{
\item{pef_code}{Integer vector of PEF species codes.}
}
\value{
Character vector of FVS species codes. Unknown codes return "OH".
}
\description{
Translates Penobscot Experimental Forest (PEF) numeric species codes
to FVS alpha codes. Regional lookup for Maine PEF inventory data.
}
\examples{
pef_to_fvs(c(1, 4, 15, 35))

}
\seealso{
Other Species Lookups: 
\code{\link{fia_to_fvs}()},
\code{\link{fvs_to_fia}()},
\code{\link{get_common_name}()},
\code{\link{get_species_table}()}
}
\concept{Species Lookups}
