% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvs_export.R
\name{set_fvs_variant}
\alias{set_fvs_variant}
\title{Set FVS Variant for the Session}
\usage{
set_fvs_variant(variant)
}
\arguments{
\item{variant}{FVS variant code. Common variants:
\itemize{
\item "NE" (Northeast)
\item "SN" (Southern)
\item "LS" (Lake States)
\item "CS" (Central States)
\item "PN" (Pacific Northwest)
\item "WC" (West Cascades)
\item "BM" (Blue Mountains)
\item "CR" (Central Rockies)
}}
}
\value{
Invisible. Sets the \code{FIAwrapper.variant} option.
}
\description{
Configures which FVS variant to use for species code translation
throughout the session. This affects all functions that use species codes.
}
\examples{
set_fvs_variant("NE")
getOption("FIAwrapper.variant")

}
\seealso{
\code{\link[=fia_to_fvs]{fia_to_fvs()}}, \code{\link[=export_fvs_treelist]{export_fvs_treelist()}}
}
