% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sqft_ac_to_sqm_ha}
\alias{sqft_ac_to_sqm_ha}
\title{Convert Square Feet per Acre to Square Meters per Hectare}
\usage{
sqft_ac_to_sqm_ha(x)
}
\arguments{
\item{x}{Numeric vector in ft^2/acre.}
}
\value{
Numeric vector in m^2/ha.
}
\description{
Convert Square Feet per Acre to Square Meters per Hectare
}
\keyword{internal}
