% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sqm_ha_to_sqft_ac}
\alias{sqm_ha_to_sqft_ac}
\title{Convert Square Meters per Hectare to Square Feet per Acre}
\usage{
sqm_ha_to_sqft_ac(x)
}
\arguments{
\item{x}{Numeric vector in m^2/ha.}
}
\value{
Numeric vector in ft^2/acre.
}
\description{
Convert Square Meters per Hectare to Square Feet per Acre
}
\keyword{internal}
