% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{summarize_growth}
\alias{summarize_growth}
\title{Summarize Growth by Plot and Period}
\usage{
summarize_growth(time_series)
}
\arguments{
\item{time_series}{A tibble from \code{\link[=build_time_series]{build_time_series()}}.}
}
\value{
A tibble per plot per period:
\describe{
\item{plot_key}{Plot identifier.}
\item{period_start, period_end}{Inventory years.}
\item{period_years}{Length of measurement period.}
\item{ba_per_acre_start, ba_per_acre_end}{BA at start/end of period.}
\item{net_ba_growth_per_acre}{Net BA change per acre per year (ft^2/ac/yr).}
\item{gross_ba_growth_per_acre}{Growth of survivors per acre per year.}
\item{mortality_ba_per_acre}{BA lost to mortality per year.}
\item{ingrowth_tpa}{New trees per acre per year entering the population.}
\item{mean_dbh_growth}{Average annual DBH increment of survivors (in/yr).}
\item{mortality_rate}{Proportion of initial TPA that died.}
}
}
\description{
Aggregates time series data to plot-level growth metrics per measurement
period. Key output for growth & yield analysis.
}
\examples{
\dontrun{
ts <- build_time_series(fia)
growth <- summarize_growth(ts)
mean(growth$net_ba_growth_per_acre, na.rm = TRUE)
}

}
\seealso{
\code{\link[=build_time_series]{build_time_series()}}, \code{\link[=get_individual_tree_history]{get_individual_tree_history()}}
}
